requirements.txt
asgiref==3.2.3
certifi==2020.4.5.2
chardet==3.0.4
dj-database-url==0.4.2
Django==2.0.2
django-bootstrap-form==3.4
django-bootstrap-pagination==1.7.1
django-cors-headers==3.1.1
django-crispy-forms==1.7.2
django-filter==1.1.0
django-heroku==0.3.1
django-taggit==1.3.0
django-widget-tweaks==1.4.1
djangorestframework==3.10.3
drf-writable-nested==0.6.0
httpie==2.1.0
idna==2.9
mysqlclient==1.4.6
Pillow==6.2.1
psycopg2==2.8.5
Pygments==2.6.1
python-decouple==3.1
pytz==2019.3
requests==2.23.0
sqlparse==0.3.0
urllib3==1.25.9
whitenoise==4.1.4